Change history screen to portrait only
- Also gets rid of auto rotate even though this may
be useful in the future. If so, the rotate functionality
can always be added back in
- This begins the evolution to improve the UX in the app.
The first step is to clean up the history screen as is into
something usable for initial launch (showing it as a modal).
In the future the history should be built into the the individual
nursing, pumping, and bottle screens. Then a new start FAB
will be added to kick off / log a feeding event that is dependent
on the tab context that you are in. The FAB can change into a an
in status bar at the bottom allowing users to stop, pause, etc.