fix: correct and improve unit display strings
For some reason, displayed fluid amount
strings would have a `x.00000...` for some
of the strings. I think this is probably due to some
rounding issues + trying to display a double with
varying precision.

This fix makes use of the built in `NumberFormatter`
for a consistent and desired output.

To solve this for the `PumpingVc`, this moves it over
to using the `SupplyAmount` struct, which is overall
move consistent with the other supply amount
implementations, more robust, and cleaner.