fix: make sure timer doesn't increment after pause
Previously the issue was that the timer wasn't steady
in incrementing due to rounding errors. This gets rid of the
"double" counter in favor of counting up by seconds using
"integers" instead. This way, we are always incrementing by
whole numbers and the timer output increments as expected
from a user's point of view.

^^ the first attempted fix to this resulted in the bug that is now being
fixed, which is that after a "pause," the timer would get set
right back to the reference date. This resulted in a timer that
didn't respect the "paused" time. In other words, after pausing
for, say, 10s, the timer would just jump those 10s when resumed.