feat: implement a _real_ in-progress view
- Functional counting timer!
- Better order for the in-progress feedings with
the most recent one added at the top of the list

This works based on having a global clock that
all timers _tick_ off of. This seems like a better pattern
than having each in-progress feeding manage their own
timer like the way the iOS app works. Probably something to
see how it works and carry over to the iOS app later :).

The double extension for time formatting was put in the
common package to support cross platform use.