feat: move timer label to use root source of truth
The timer labels now depend on the in-progress
feeding model to show the time, count up, pause, and stop.
This is more reliable in that we are no longer duplicating state
(which interestingly enough used to be the case, but for
some reason past me had the "bright idea" to isolate the
timer label from the single source of truth...oh well :)).

Impact:
- Improves reaction to changes to in-progress feedings from the
Apple Watch app
- Fixes a bug where the paused time was not correctly accounted for
in the timer UI even though the model was correct

Also, update tests.