feat: account for paused feeding
The watch will now properly account for
paused time after a period of being paused.

Right now the UI just says "PAUSED".
In the future we can build out a blinking
paused timer label like the main app, but this
is okay for now.

Heavily based off of the main app's
implementation of the feeding struct. The
two different feeding structs (watch and iOS)
should be combined in the future.

I am not too crazy about duplicating the state
of in-progress feedings here, but due to the
difficulty in watch<>iOS app communication,
this is okay for now.

A resume action was formally built in too.