fix: stop main app communication
Unfortunately, when sending a feeding
from the watch to the phone, the communication
would "boomerang" back to the watch and duplicate
what was just sent (e.g., duplicate feedings would show
up on the watch). This commit breaks that communication
for the time being.

This was an oversight on my part :(. That
is okay though, this exposed how fragile
things were getting and that we should probably
be using the context for the main app's communication.

By using the context, that will also give the watch the
potential of receiving all the logged feedings!