feat: add tests for main app Feeding model entity
This will help support refactoring to add
codable conformance.

Also, fixes the full paused time to make sure
we account for an ended feeding. If not, the paused
time could previously go past the ended time to
give an inflated amount of paused time, which is wrong. This wasn't an
issue previously because management of these properties was handled
correctly (i.e., lastPausedDate and endDate were mutually exclusive).

We should, however, also enforce in code that you can't set a
lastPausedDate at a later time than the endDate.