fix: remove duplicate newly started feeding
Due to the async nature of sending the communication
there is a chance that the watch will end up showing
two feedings when starting a feeding.
i.e., the context updating shows the new feeding
and then the watch can sometimes add a new feeding on
top of that if the response form the send message comes
in after.