feat: improve synchronization of app state context
When starting the watch app, it will ask the main app for
the full app state in context form. The main app will then
send back the full state including feedings and logged in
status.

This greatly improves the initial starting up experience
of the watch app! Also, this sets the stage for the watch to
have a simple UI mechanism to refresh the full state from the
main app!!