feat: remove new feeding actions / reducer
no longer needed since watch state is handled by
getting updates from the context. And, SwiftUI is great
at managing the UI updates due to context changes!

This also included the removal of the communication
failure state/action since that was a direct side effect
of the new feeding reducer.