feat: make use of loading spinner
Show loading spinner on feeding actions
(start, pause, resume, stop).

The loading spinner will not be double shown
with dialogs since they are all mutually exclusive in
being shown due to the if / if else chain for showing popups
like this. Due to this, combining the dialog and loading booleans
into an enum might be a better way to better lock down the
mutual exclusivity.