feat: remove implicit sending on main thread
This removes the implicit _all sends are handled on the
main thread_ in favor of local thread management.

The reducer effects that dispatch to the main thread have
the potential to be better addressed with a tweak to the effect
type that allows chaining a `receive(on:)` to specify the receiving
thread.

Also, the fact that some of the sends need to be wrapped in
a dispatch to main could be a code smell in that the delegate calls
should be directly sent to the reducer and resulting operations be preformed as
a side effect, which will have better, more explicit, threading management.

^^ something to address in the future, but the current implementation is already
better in that the treads need to be thought about on a case by case basis.