fix: prevent false comm error
This happens because this call
to get the context was too early in the lifecycle
to make the context request, and would fail every time.

This was being hidden by implicit dispatch all to main
that was recently removed, so this bug popped up, which was
already there but hidden.