fix: plot history graph bars using event start date
After changing the summary to use the start
date, the graph needed to also be switched over too.
This corrects that.

This disconnect between the history stats and bar graph
is a source of weakness. As a result, this dependency could
use some tests or an effort to codify the dependency so this
doesn't happen again.